requirements.txt
Flask==2.0.1
Keras-Preprocessing==1.1.2
numpy
requests==2.22.0
requests-oauthlib==1.3.0
tensorflow-cpu==2.10.0